Move task workers to worker modules
